Move QuickActionsOptions to runtimes-types and consume it in chat-ui-types
