+++ runtimes/protocol/chat.ts
@@ -31,34 +31,6 @@ import {
     TAB_REMOVE_NOTIFICATION_METHOD,
 } from './lsp'
 
-/**
- * Configuration object for chat quick action.
- */
-export interface QuickActionCommand {
-    command: string
-    disabled?: boolean
-    description?: string
-    placeholder?: string
-}
-
-/**
- * Configuration object for registering chat quick actions groups.
- */
-export interface QuickActionCommandGroup {
-    groupName?: string
-    commands: QuickActionCommand[]
-}
-
-/**
- * Registration options for a Chat QuickActionRequest.
- */
-export interface QuickActionsOptions {
-    /**
-     * The chat quick actions groupd and commands to be executed on server.
-     */
-    quickActionsCommandGroups: QuickActionCommandGroup[]
-}
-
 export interface ChatRequest extends ChatParams {
     partialResultToken?: ProgressToken
 }
