+++ types/chat.ts
@@ -86,6 +86,34 @@ export interface ChatResult {
 export type EndChatParams = { tabId: string }
 export type EndChatResult = boolean
 
+/**
+ * Configuration object for chat quick action.
+ */
+export interface QuickActionCommand {
+    command: string
+    disabled?: boolean
+    description?: string
+    placeholder?: string
+}
+
+/**
+ * Configuration object for registering chat quick actions groups.
+ */
+export interface QuickActionCommandGroup {
+    groupName?: string
+    commands: QuickActionCommand[]
+}
+
+/**
+ * Registration options for a Chat QuickActionRequest.
+ */
+export interface QuickActionsOptions {
+    /**
+     * The chat quick actions groups and commands to be executed on server.
+     */
+    quickActionsCommandGroups: QuickActionCommandGroup[]
+}
+
 export interface QuickActionParams extends PartialResultParams {
     tabId: string
     quickAction: string
